SSH fixes for non-blocking RunLoop & SSH Channel EOF handling.
- When running fast requests through a tunnel, we experienced that some would
get stuck in the RunLoop until an event would trigger it to wake up. This would
happen even if we have a timeout.
- LibSSH now performs all the requests in a non-blocking manner.
- The read loops have been simplified as it cannot happen anymore that the state
of the streams changes in the middle of an operation (like getting cancelled).
- Tunnels would close with an EOF callback the streams from the other side, so
the stream needs to publish that event.
- SFTP is the only one that requires a channel to be blocking, except when
dealing with file streams. We have fixed those.