Local attributes as part of the Cache and Item Reference
- We use the local Filesystem to elaborate the cache and create the
representation of what got coordinated between local and remote.
- Next we need to elaborate and offer the state based on this local and remote
parts. BlinkItemReference then will be the one offering the FileItem interface,
while being responsible to handle the BlinkFiles interface.
- There is a chance we may not need the FilesList after all, because the
reference is already aggregating this information. We may need it in the future
though, when passing changes. We will need the original list.