Constraint LocalFiles memory usage while reading
When a Dispatch channel reads from a file up to INTMAX, it tries to fit as much
of that file in memory as it possibly can. This does not play very well with the
extension, that requires to limit memory usage. Low and high watermarks seem to
not have any effect. It is unclear when memory is released.

We are changing the algorithm when demand is not unbounded, for the WriterTo
case. We are reading in a reentry way and that seems to keep memory bounded.