FileProvider issues traversing paths with accents. Fixes #1321
Although SFTP servers usually work with UTF8 charsets, the same name attribute
with an accent may have different representations. The FileProvider was using a
URL which performed a character translation that did not then match the UTF8 one
coming from the server.

The fix is to not generate a BlinkIdentifier from a URL, as it is not a valid
byte representation, and only use the one coming from the Server.